+++ test/completion/init.lua
@@ -905,7 +905,7 @@ local ast = 1
 local t = 'as$'
 local ask = 1
 ]]
-(nil)
+(EXISTS)
 
 TEST [[
 local add