+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.10.0
 * `NEW` workspace: supports `.dll`(`.so`) in `require`
 * `CHG` supports `~` in command line
+* `CHG` completion: improve workspace words
 * `FIX` [#339](https://github.com/sumneko/lua-language-server/issues/339)
 
 ## 1.9.0