+++ changelog.md
@@ -1,5 +1,8 @@
 # changelog
 
+## 1.11.0
+* `FIX` runtime errors
+
 ## 1.10.0
 `2021-1-4`
 * `NEW` workspace: supports `.dll`(`.so`) in `require`